fix MRPD

+++ file_comparison/stats.py
@@ -88,7 +88,8 @@ def mean_percentage_error(origin:np.ndarray, new:np.ndarray):
 # TO FIX
 def mean_relative_percentage_difference(origin:np.ndarray, new:np.ndarray):
 
-    core = np.divide (np.abs(origin.astype(np.float64) - new.astype(np.float64), dtype=np.float64), ((origin.astype(np.float64) + new.astype(np.float64)).astype(np.float64)/2.), out=np.full_like(origin, np.nan), where=(((origin.astype(np.float64) + new.astype(np.float64))/2.)!=0))
+    core = np.divide (np.abs(origin.astype(np.float64) - new.astype(np.float64), dtype=np.float64).astype(np.float64), ((origin.astype(np.float64) + new.astype(np.float64)).astype(np.float64)/2.).astype(np.float64), out=np.full_like(origin, np.nan), where=(((origin.astype(np.float64) + new.astype(np.float64))/2.)!=0))
+    
     return np.nanmean (core)*100.
